+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 1.8.1 (Nov 19, 2023)
+
+### Fixes
+- Resolved an issue with linkedConditions not behaving correctly when the array includes both strings and arrays.
+- Fixed a bug where linkedConditions were not functioning correctly when reverseConditions were enabled, along with other conditions.
+
+### Changes
+- Changed name and description.
+- Changed the order of remove conditions in config.
+
 ## Version 1.8 (Nov 14, 2023)
 
 ### Features